backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
azure-core==1.38.2
    # via
    #   azure-cosmos
    #   azure-storage-blob
    #   azure-storage-queue
azure-cosmos==4.15.0
    # via embedding-model-platform (pyproject.toml)
azure-storage-blob==12.28.0
    # via embedding-model-platform (pyproject.toml)
azure-storage-queue==12.15.0
    # via embedding-model-platform (pyproject.toml)
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
cryptography==46.0.5
    # via
    #   azure-storage-blob
    #   azure-storage-queue
distro==1.9.0
    # via openai
fastapi==0.135.1
    # via embedding-model-platform (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
isodate==0.7.2
    # via
    #   azure-storage-blob
    #   azure-storage-queue
jiter==0.13.0
    # via openai
openai==2.24.0
    # via embedding-model-platform (pyproject.toml)
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   embedding-model-platform (pyproject.toml)
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via embedding-model-platform (pyproject.toml)
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.22
    # via embedding-model-platform (pyproject.toml)
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via azure-core
sniffio==1.3.1
    # via openai
starlette==0.52.1
    # via fastapi
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   azure-core
    #   azure-cosmos
    #   azure-storage-blob
    #   azure-storage-queue
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via embedding-model-platform (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
